Fix Check*Exist db functions.
Because MySQL doesn't guarantee the order of rows when
using UNION ALL, we have to ensure that the rows are returned
in the expected order, which is that of the list of parameters.
Because we haven't been able to make it fail in tests or real life,
there is no tests checking for this situation.