Update dependencies.
In particular, the mysql driver has an issue where it closes
the connection if using TCP and gets an invalid connection error.
The error is "unexpected EOF", and can be work around by setting
the number of idle connections to 0. SetMaxIdleConns(0).
But updating the driver should also fix the bug.
https://github.com/go-sql-driver/mysql/issues/674