Fix two bugs: Source.Prepare had a race condition, readIncomingSequentially a typo.
The method Source.Prepare expected to run its goroutine on the error channel of the
stage, but was not guaranteed if it took a bit longer than the replacement of the
error channel happening in auto resume.
The method readIncomingSequentially was always reading from the stage's incoming
channel slice, instead of the temporary copy that was dynamically altered.